Fixed stacking of link,image,youtube and email
Solved an issue where if you highlight a link you have inserted and go to insert another link in its place using the drop down in WYSIWYG mode, this will stop the element stacking. Example current output:

```
<a href="http://1stlink.com">
    <a href="http://2ndlink.com">DESCRIPTION</a>
</a>
```

In addition to this, the previous link you have highlighted to change will display the link details in the drop down allowing you to edit the previous link instead of having to retype everything again. The same goes for modifying email, images and YouTube Videos.

I have used https:// for YouTube as http will be classed as insecure in 2017.