Fix going back in the browser causes content to be lost
If you use the back button in your browser then go forward, all the content is lost.  This is because the information is not sent back to the original prior to unloading the page.  This change simply attaches a event to the unload for window and triggers the base update to occur.  The browser handles the rest if it supports remembering what was entered into the textarea.

Signed-off-by: jdarwood007 <unmonitored+github@sleepycode.com>