Update textarea value on pagehide so back button and suspend don't loose value (#857)
* Fix going back in the browser causes content to be lost

If you use the back button in your browser then go forward, all the content is lost.  This is because the information is not sent back to the original prior to unloading the page.  This change simply attaches a event to the unload for window and triggers the base update to occur.  The browser handles the rest if it supports remembering what was entered into the textarea.

Signed-off-by: jdarwood007 <unmonitored+github@sleepycode.com>

* Use page events instead of unload for back fix.

Using the unload event can cause the browser not to use the bfcache:
https://web.dev/bfcache/
https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Releases/1.5/Using_Firefox_1.5_caching

Co-authored-by: jdarwood007 <unmonitored+github@sleepycode.com>