Fix emoticons not being converted and tidy code
Emoticons should still be converted if they find their way into a code
tag as this is the current behaviour.

vChildren will either be an array or undefined so no need to check type.

hasCodeParent is now always passed so make that second argument.

Adding text node check to if statement removes need for early return
nodeType check.